+++ index.html
@@ -231,6 +231,6 @@
   </template>
 
   <!-- Подключаем файл JS -->
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
