+++ js/get-random-positive-integer.js
@@ -1,0 +1,14 @@
+// 1. Типовая функция - вычисляем случайное целое число из заданного диапазона.
+function getRandomIntInclusive(min, max) {
+  if (min >= 0 && max >= 0 && max > min) {
+    min = Math.ceil(min); // округление вверх. Округляет аргумент до ближайшего большего целого.
+    max = Math.floor(max); // округление вниз. Округляет аргумент до ближайшего меньшего целого.
+    const integer = Math.floor(Math.random() * (max - min + 1)) + min; //Максимум и минимум включаются.
+    return integer;
+  } else if (min < 0 || max < 0) {
+    //Выводим сообщение об обшибке - ('В диапазоне должны использоваться только положительные числа!');
+  } else if (max <= min) {
+    //Выводим сообщение об обшибке - ('В диапазоне максимальное значение должно быть больше минимального!');
+  }
+}
+export {getRandomIntInclusive};
