+++ js/variables.js
@@ -1,0 +1,48 @@
+// Модуль для хранения переменных
+
+const USER_NUMBER_MIN = 1;
+const USER_NUMBER_MAX = 8;
+const PRICE_VALUE_MIN = 1;
+const PRICE_VALUE_MAX = 50;
+const TYPE_OF_APARTMENT = ['Palace', 'Flat', 'House', 'Bungalow', 'Hotel'];
+const ROOMS_VALUE_MIN = 1;
+const ROOMS_VALUE_MAX = 5;
+const GUEST_VALUE_MIN = 1;
+const GUEST_VALUE_MAX = 10;
+const CHECK_IN_TIME = ['12:00', '13:00', '14:00'];
+const CHECK_OUT_TIME = ['12:00', '13:00', '14:00'];
+const HOTEL_FEATURES = ['Wifi', 'Dishwasher', 'Parking', 'Washer', 'Elevator', 'Conditioner'];
+const REAL_ESTATE_PHOTO = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+const LONGITUDE_VALUE_MIN = 35.65000;
+const LONGITUDE_VALUE_MAX = 35.70000;
+const LATITUDE_VALUE_MIN = 139.70000;
+const LATITUDE_VALUE_MAX = 139.80000;
+const COORDS_ACCURACY = 5;
+const NUMBER_OF_ARRAYS = 10;
+
+
+// экспортируем все обьявленые переменные
+export {
+  USER_NUMBER_MIN,
+  USER_NUMBER_MAX,
+  PRICE_VALUE_MIN,
+  PRICE_VALUE_MAX,
+  TYPE_OF_APARTMENT,
+  ROOMS_VALUE_MIN,
+  ROOMS_VALUE_MAX,
+  GUEST_VALUE_MIN,
+  GUEST_VALUE_MAX,
+  CHECK_IN_TIME,
+  CHECK_OUT_TIME,
+  HOTEL_FEATURES,
+  REAL_ESTATE_PHOTO,
+  LONGITUDE_VALUE_MIN,
+  LONGITUDE_VALUE_MAX,
+  LATITUDE_VALUE_MIN,
+  LATITUDE_VALUE_MAX,
+  COORDS_ACCURACY,
+  NUMBER_OF_ARRAYS
+};
